backend/holdings/serializers.py
from rest_framework import serializers

from .models import Holding


class HoldingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Holding
        fields = ('run', 'quantity')

