backend/requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
black==19.3b0
Click==7.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.3
drf-nested-routers==0.91
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
python-decouple==3.1
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
wcwidth==0.1.7
wrapt==1.11.2
